sam-geo-information/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.sam_geo_information.metadata]
type = "agent"

[project]
name = "sam_geo_information"
version = "0.1.0"
authors = [
  { name="SolaceLabs", email="solacelabs@solace.com" },
]
description = "A plugin that provides comprehensive geographic information services including location, timezone, and weather data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests==2.32.5",
    "PyYAML==6.0.2"
]

[tool.hatch.build.targets.wheel]
packages = ["src/sam_geo_information"]
src-path = "src"

[tool.hatch.build.targets.wheel.force-include]
"src/sam_geo_information" = "sam_geo_information/"
"config.yaml" = "sam_geo_information/config.yaml"
"README.md" = "sam_geo_information/README.md"
"pyproject.toml" = "sam_geo_information/pyproject.toml"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--tb=short --strict-markers --disable-warnings -p pytest_asyncio"
markers = []

[tool.hatch.envs.hatch-test]
dependencies = [
  "pytest-asyncio",
  "pytest>=8.0.0",
  "pytest-mock>=3.0.0",
  "pytest-cov>=4.0.0",
  "pytest-xdist>=3.5.0",
  "sam-test-infrastructure @ git+https://github.com/SolaceLabs/solace-agent-mesh#subdirectory=tests/sam-test-infrastructure"
]
